+++ py/legacyccds/legacy-zeropoints.py
@@ -1,6 +1,37 @@
 #!/usr/bin/env python
 
-"""Generate a legacypipe-compatible CCD-level zeropoints file for a given set of
+"""
+RUN
+=== 
+1) generate file list of cpimages, e.g. for everything mzls
+find /project/projectdirs/cosmo/staging/mosaicz/MZLS_CP/CP*v2/k4m*ooi*.fits.fz > mosaic_allcp.txt
+2) use batch script "submit_zpts.sh" to run legacy-zeropoints.py
+-- scaling is good with 10,000+ cores
+   -- key is Yu Feng's bcast, we made tar.gz files for all the NERSC HPCPorts modules and Yu's bcast efficiently copies them to ram on every compute node for fast python startup
+   -- the directory containing the tar.gz files is /global/homes/k/kaylanb/repos/yu-bcase
+      also on kaylans tape: bcast_hpcp.tar
+-- debug queue gets all zeropoints done < 30 min
+-- set SBATCH -N to be as many nodes as will give you mpi tasks = nodes*cores_per_nodes ~ number of cp images
+-- ouput ALL plots with --verboseplots, ie Moffat PSF fits to 20 brightest stars for FWHM
+3) Make a file (e.g. zpt_files.txt) listing all the zeropoint files you just made (not includeing the -star.fits ones), then 
+    a) compare legacy zeropoints to Arjuns 
+        run from loggin node: 
+        python legacy-zeropoints.py --image_list zpt_files.txt --compare2arjun
+    b) gather all zeropoint files into one fits table
+        run from loggin node
+            python legacy-zeropoints-gather.py --file_list zpt_files.txt --nproc 1 --outname gathered_zpts.fits
+        OR if that takes too long, runs out of memory, etc, run with mpi tasks
+            uncomment relavent lines of submit_zpts.sh, comment out running legacy-zeropoints.py
+            <= 50 mpi tasks is fine
+            sbatch submit_zpts.sh 
+
+UNITS
+=====
+match Arjun's decstat,moststat,bokstat.pro 
+
+DOC
+===
+Generate a legacypipe-compatible CCD-level zeropoints file for a given set of
 (reduced) BASS, MzLS, or DECaLS imaging.
 
 This script borrows liberally from code written by Ian, Kaylan, Dustin, David
@@ -37,8 +68,15 @@ Proposed changes to the -ccds.fits file used by legacypipe:
    CCDs file should be directly calculated and stored here?
  * Are ccdnum and image_hdu redundant?
 
+REVISION HISTORY:
+    13-Sept-2016 J. Moustakas
+    15-Dec-2016  K. Burleigh
 """
 from __future__ import division, print_function
+if __name__ == '__main__':
+    import matplotlib
+    matplotlib.use('Agg')
+import matplotlib.pyplot as plt
 
 import os
 import pdb
@@ -46,17 +84,34 @@ import argparse
 
 import numpy as np
 from glob import glob
+from scipy.optimize import curve_fit
+from scipy.stats import sigmaclip
+from scipy.ndimage.filters import median_filter
 
 import fitsio
 from astropy.table import Table, vstack
+from astropy import units
+from astropy.coordinates import SkyCoord
 from astrometry.util.starutil_numpy import hmsstring2ra, dmsstring2dec
 from astrometry.util.ttime import Time
 import datetime
-import matplotlib.pyplot as plt
 import sys
 
+from photutils import (CircularAperture, CircularAnnulus,
+                       aperture_photometry, daofind)
+
+from astrometry.util.fits import fits_table, merge_tables
+from astrometry.util.util import wcs_pv2sip_hdr
+from astrometry.libkd.spherematch import match_radec
+from astrometry.libkd.spherematch import match_xy
+
+from tractor.splinesky import SplineSky
+
+from legacyanalysis.ps1cat import ps1cat
+
 ######## 
-## Ted's
+# stdouterr_redirected() is from Ted Kisner
+# Every mpi task (zeropoint file) gets its own stdout file
 import time
 from contextlib import contextmanager
 
@@ -127,7 +182,6 @@ def stdouterr_redirected(to=os.devnull, comm=None):
             sys.stderr.flush()
             
     return
-##############
 
 def ptime(text,t0):
     tnow=Time()
@@ -144,6 +198,19 @@ def read_lines(fn):
 def dobash(cmd):
     print('UNIX cmd: %s' % cmd)
     if os.system(cmd): raise ValueError
+
+
+def extra_ccd_keys(camera='decam'):
+    '''Returns list of camera-specific keywords for the ccd table'''
+    if camera == 'decam':
+        keys= [('ccdzpta', '>f4'), ('ccdzptb','>f4'), ('ccdnmatcha', '>i2'), ('ccdnmatchb', '>i2'),\
+               ('temp', '>f4')]
+    elif camera == 'mosaic':
+        keys=[]
+    elif camera == '90prime':
+        keys=[('ccdzpt1', '>f4'), ('ccdzpt2','>f4'), ('ccdzpt3', '>f4'), ('ccdzpt4','>f4'),\
+              ('ccdnmatcha', '>i2'), ('ccdnmatch2', '>i2'), ('ccdnmatch3', '>i2'), ('ccdnmatch4', '>i2')]
+    return keys 
 
 def _ccds_table(camera='decam'):
     '''Initialize the output CCDs table.  See decstat.pro and merge-zeropoints.py
@@ -168,7 +235,7 @@ def _ccds_table(camera='decam'):
         ('ha', 'S13'),             # hour angle (from header)
         ('airmass', '>f4'),        # airmass (from header)
         #('seeing', '>f4'),        # seeing estimate (from header, arcsec)
-        #('fwhm', '>f4'),          # FWHM (pixels)
+        ('fwhm', '>f4'),          # FWHM (pixels)
         #('arawgain', '>f4'),       
         ('gain', '>f4'),           # average gain (camera-specific, e/ADU) -- remove?
         #('avsky', '>f4'),         # average sky value from CP (from header, ADU) -- remove?
@@ -206,34 +273,35 @@ def _ccds_table(camera='decam'):
         ]
 
     # Add camera-specific keywords to the output table.
-    if camera == 'decam':
-        cols.extend([('ccdzpta', '>f4'), ('ccdzptb','>f4'), ('ccdnmatcha', '>i2'), ('ccdnmatchb', '>i2'),
-                     ('temp', '>f4')])
-    elif camera == 'mosaic':
-        pass
-    elif camera == '90prime':
-        cols.extend([('ccdzpt1', '>f4'), ('ccdzpt2','>f4'), ('ccdzpt3', '>f4'), ('ccdzpt4','>f4'),
-                     ('ccdnmatcha', '>i2'), ('ccdnmatch2', '>i2'), ('ccdnmatch3', '>i2'), ('ccdnmatch4', '>i2')])
-        
+    #cols.extend( extra_ccd_keys(camera=camera) )
+    
     ccds = Table(np.zeros(1, dtype=cols))
     return ccds
 
+     
 def _stars_table(nstars=1):
     '''Initialize the stars table, which will contain information on all the stars
        detected on the CCD, including the PS1 photometry.
 
     '''
-    cols = [('expid', 'S16'), ('filter', 'S1'), ('amplifier', 'i2'), ('x', 'f4'), ('y', 'f4'),
-            ('ra', 'f8'), ('dec', 'f8'), ('fwhm', 'f4'), ('apmag', 'f4'),
-            ('gaia_ra', 'f8'), ('gaia_dec', 'f8'), ('ps1_mag', 'f4'), ('ps1_gicolor', 'f4')]
-            #('ps1_ra', 'f8'), ('ps1_dec', 'f8'), ('ps1_mag', 'f4'), ('ps1_gicolor', 'f4')]
+    cols = [('image_filename', 'S65'),('expid', 'S16'), ('filter', 'S1'),('nmatch', '>i2'), 
+            ('amplifier', 'i2'), ('x', 'f4'), ('y', 'f4'),
+            ('ra', 'f8'), ('dec', 'f8'), ('apmag', 'f4'),
+            ('radiff', 'f8'), ('decdiff', 'f8'),('radiff_ps1', 'f8'), ('decdiff_ps1', 'f8'),
+            ('gaia_ra', 'f8'), ('gaia_dec', 'f8'), ('ps1_mag', 'f4'), ('ps1_gicolor', 'f4'),
+            ('gaia_g','f8'),('ps1_g','f8'),('ps1_r','f8'),('ps1_i','f8'),('ps1_z','f8')]
     stars = Table(np.zeros(nstars, dtype=cols))
-
     return stars
+
+def getrms(x):
+    return np.sqrt( np.mean( np.power(x,2) ) )
+
+def moffatPSF(x, a, r0, beta):
+    return a*(1. + (x/r0)**2)**(-beta)
 
 class Measurer(object):
     def __init__(self, fn, ext, aprad=3.5, skyrad_inner=7.0, skyrad_outer=10.0,
-                 sky_global=False, calibrate=False, only_ccd_centers=False):
+                 sky_global=False, calibrate=False,**kwargs):
         '''This is the work-horse class which operates on a given image regardless of
         its origin (decam, mosaic, 90prime).
 
@@ -246,14 +314,16 @@ class Measurer(object):
         Sky annulus radius in arcsec
 
         '''
-        from astrometry.util.util import wcs_pv2sip_hdr
+        # Set extra kwargs
+        self.zptsfile= kwargs.get('zptsfile')
+        self.prefix= kwargs.get('prefix')
+        self.verboseplots= kwargs.get('verboseplots')
         
         self.fn = fn
         self.ext = ext
 
         self.sky_global = sky_global
         self.calibrate = calibrate
-        self.only_ccd_centers=only_ccd_centers
         
         self.aprad = aprad
         self.skyrad = (skyrad_inner, skyrad_outer)
@@ -261,9 +331,8 @@ class Measurer(object):
         # Set the nominal detection FWHM (in pixels) and detection threshold.
         self.nominal_fwhm = 5.0 # [pixels]
         self.det_thresh = 10    # [S/N] - used to be 20
-        self.stampradius = 15   # stamp radius around each star [pixels]
-
-        self.matchradius = 1. #2.0  # search radius for finding matching PS1 stars [arcsec]
+        #self.stampradius = 15   # tractor fitting no longer done, stamp radius around each star [pixels]
+        self.matchradius = 1. # Matching to PS1 [arcsec]
 
         # Read the primary header and the header for this extension.
         self.primhdr = fitsio.read_header(fn, ext=0)
@@ -276,28 +345,24 @@ class Measurer(object):
         self.mjd_obs = self.primhdr['MJD-OBS']
         self.airmass = self.primhdr['AIRMASS']
         self.ha = self.primhdr['HA']
-        #self.seeing = self.primhdr['SEEING']
-
-        if 'EXPNUM' in self.hdr: # temporary hack!
-            self.expnum = self.hdr['EXPNUM']
+        
+        # FIX ME!, gets unique id for mosaic but not 90prime
+        if 'EXPNUM' in self.primhdr: 
+            self.expnum = self.primhdr['EXPNUM']
         else:
+            print('WARNING! no EXPNUM in %s' % self.fn)
             self.expnum = np.int32(os.path.basename(self.fn)[11:17])
 
-        self.ccdname = self.hdr['EXTNAME'].strip().upper()
+        self.ccdname = self.hdr['EXTNAME'].strip()
         self.image_hdu = np.int(self.hdr['CCDNUM'])
-        #self.ccdnum = self.hdr['CCDNUM']
-        #self.image_hdu = np.int(self.ccdnum)
 
         self.expid = '{:08d}-{}'.format(self.expnum, self.ccdname)
-        self.band = self.get_band()
 
         self.object = self.primhdr['OBJECT']
 
-        self.wcs = wcs_pv2sip_hdr(self.hdr) # PV distortion
-        self.pixscale = self.wcs.pixel_scale()
-
-        # Eventually we would like FWHM to not come from SExtractor.
-        #self.fwhm = 2.35 * self.primhdr['SEEING'] / self.pixscale  # [FWHM, pixels]
+        self.wcs = self.get_wcs()
+        # Pixscale is assumed CONSTANT! per camera
+        #self.pixscale = self.wcs.pixel_scale()
 
     def zeropoint(self, band):
         return self.zp0[band]
@@ -313,8 +378,6 @@ class Measurer(object):
         mean(goodpix) +- nsigma * sigma
 
         '''
-        from scipy.stats import sigmaclip
-        
         goodpix, lo, hi = sigmaclip(arr, low=nsigma, high=nsigma)
         meanval = np.mean(goodpix)
         sigma = (meanval - lo) / nsigma
@@ -322,7 +385,6 @@ class Measurer(object):
 
     def get_sky_and_sigma(self, img):
         # Spline sky model to handle (?) ghost / pupil?
-        from tractor.splinesky import SplineSky
 
         #sky, sig1 = self.sensible_sigmaclip(img[1500:2500, 500:1000])
 
@@ -334,7 +396,6 @@ class Measurer(object):
         return skyimg, sig1
 
     def remove_sky_gradients(self, img):
-        from scipy.ndimage.filters import median_filter
         # Ugly removal of sky gradients by subtracting median in first x and then y
         H,W = img.shape
         meds = np.array([np.median(img[:,i]) for i in range(W)])
@@ -345,7 +406,6 @@ class Measurer(object):
         img -= meds[:,np.newaxis]
 
     def match_ps1_stars(self, px, py, fullx, fully, radius, stars):
-        from astrometry.libkd.spherematch import match_xy
         #print('Matching', len(px), 'PS1 and', len(fullx), 'detected stars with radius', radius)
         I,J,d = match_xy(px, py, fullx, fully, radius)
         #print(len(I), 'matches')
@@ -423,21 +483,23 @@ class Measurer(object):
         
         return np.array(fwhms)
 
+    def isolated_radec(self,ra,dec,nn=2,minsep=1./3600):
+        '''return indices of ra,dec for which the ra,dec points are 
+        AT LEAST a distance minsep away from their nearest neighbor point'''
+        cat1 = SkyCoord(ra=ra*units.degree, dec=dec*units.degree)
+        cat2 = SkyCoord(ra=ra*units.degree, dec=dec*units.degree)
+        idx, d2d, d3d = cat1.match_to_catalog_3d(cat2,nthneighbor=nn)
+        b= np.array(d2d) >= minsep
+        return b
+
     def run(self):
         t0= Time()
-        from scipy.stats import sigmaclip
-        from legacyanalysis.ps1cat import ps1cat
-        from astrometry.libkd.spherematch import match_radec
-        from photutils import (CircularAperture, CircularAnnulus,
-                               aperture_photometry, daofind)
         t0= ptime('import-statements-in-measure.run',t0)
 
         # Read the image and header.
         print('Todo: read the ivar image')
         img, hdr = self.read_image()
-        #fits=fitsio.FITS(fn,mode='r',clobber=False,lower=True)
-        #hdr= fits[0].read_header()
-        #img= fits[ext].read()
+        bitmask = self.read_bitmask()
 
         # Initialize and begin populating the output CCDs table.
         ccds = _ccds_table(self.camera)
@@ -447,7 +509,6 @@ class Measurer(object):
         ccds['camera'] = self.camera
         ccds['expnum'] = self.expnum
         ccds['ccdname'] = self.ccdname
-        #ccds['ccdnum'] = self.ccdnum
         ccds['expid'] = self.expid
         ccds['object'] = self.object
         ccds['propid'] = self.propid
@@ -460,7 +521,6 @@ class Measurer(object):
         ccds['dec_bore'] = self.dec_bore
         ccds['ha'] = self.ha
         ccds['airmass'] = self.airmass
-        #ccds['fwhm'] = self.fwhm
         ccds['gain'] = self.gain
         ccds['pixscale'] = self.pixscale
 
@@ -482,9 +542,6 @@ class Measurer(object):
         ccds['ra'] = ccdra   # [degree]
         ccds['dec'] = ccddec # [degree]
         t0= ptime('read-image-getheader-info',t0)
-        if self.only_ccd_centers:
-            print('Stopping early, only_ccd_centers=',self.only_ccd_centers)
-            return ccds, _stars_table()
 
         # Measure the sky brightness and (sky) noise level.  Need to capture
         # negative sky.
@@ -496,14 +553,14 @@ class Measurer(object):
         sky, sig1 = self.get_sky_and_sigma(img)
         sky1 = np.median(sky)
         skybr = zp0 - 2.5*np.log10(sky1 / self.pixscale / self.pixscale / exptime)
-        #skybr = zp0 - 2.5*np.log10(sky1 / self.pixscale / self.pixscale)
-
         print('  Sky brightness: {:.3f} mag/arcsec^2'.format(skybr))
         print('  Fiducial:       {:.3f} mag/arcsec^2'.format(sky0))
 
         ccds['skyrms'] = sig1    # [electron/pix]
         ccds['skycounts'] = sky1 # [electron/pix]
         ccds['skymag'] = skybr   # [mag/arcsec^2]
+        ccds['skyrms'] /= exptime   
+        ccds['skycounts'] /= exptime 
         t0= ptime('measure-sky',t0)
 
         # Detect stars on the image.  
@@ -526,7 +583,7 @@ class Measurer(object):
 
         # Do aperture photometry in a fixed aperture but using either local (in
         # an annulus around each star) or global sky-subtraction.
-        print('Performing aperture photometry -- need to include the mask!')
+        print('Performing aperture photometry')
 
         ap = CircularAperture((obj['xcentroid'], obj['ycentroid']), self.aprad / self.pixscale)
         if self.sky_global:
@@ -539,124 +596,201 @@ class Measurer(object):
             apphot = aperture_photometry(img, ap)
             skyphot = aperture_photometry(img, skyap)
             apflux = apphot['aperture_sum'] - skyphot['aperture_sum'] / skyap.area() * ap.area()
-        t0= ptime('aperture-photometry',t0)
+        # Use Bitmask, remove stars if any bitmask within 5 pixels
+        bit_ap = CircularAperture((obj['xcentroid'], obj['ycentroid']), 5.)
+        bit_phot = aperture_photometry(bitmask, bit_ap)
+        bit_flux = bit_phot['aperture_sum'] 
+        # No stars within 13''
+        minsep = 13. #arcsec
+        objra, objdec = self.wcs.pixelxy2radec(obj['xcentroid']+1, obj['ycentroid']+1)
+        b_isolated= self.isolated_radec(objra,objdec,nn=2,minsep=minsep/3600.)
+        # Aperture mags
+        apmags= - 2.5 * np.log10(apflux.data) + zp0 + 2.5 * np.log10(exptime)
 
-        # Remove stars with negative flux (or very large photometric uncertainties).
-        istar = np.where(apflux > 0)[0]
+        # Good stars following IDL codes
+        # We are ignoring aperature errors though
+        minsep_px = minsep/self.pixscale
+        istar = np.where((apflux > 0)*\
+                         (bit_flux == 0)*\
+                         (b_isolated == True)*\
+                         (apmags > 12.)*\
+                         (apmags < 30.)*\
+                         (obj['xcentroid'] > minsep_px)*\
+                         (obj['xcentroid'] < img.shape[0]-minsep_px)*\
+                         (obj['ycentroid'] > minsep_px)*\
+                         (obj['ycentroid'] < img.shape[1]-minsep_px))[0] #KJB
+
         if len(istar) == 0:
-            print('All stars have negative aperture photometry!')
+            print('FAIL: All stars have negative aperture photometry AND/OR contain masked pixels!')
             return ccds, _stars_table()
         obj = obj[istar]
+        objra, objdec = self.wcs.pixelxy2radec(obj['xcentroid']+1, obj['ycentroid']+1)
         apflux = apflux[istar].data
         ccds['nstar'] = len(istar)
+        t0= ptime('aperture-photometry',t0)
 
-        # Now match against (good) PS1 stars with magnitudes between 15 and 22.
-        ps1 = ps1cat(ccdwcs=self.wcs).get_stars(magrange=(15, 22))
-        good = np.where((ps1.nmag_ok[:, 0] > 0)*(ps1.nmag_ok[:, 1] > 0)*(ps1.nmag_ok[:, 2] > 0))[0]
+        # FWHM: fit moffat profile to 20 brightest stars
+        # annuli 0.5'' --> 3.5''
+        radii = np.linspace(0.5/self.pixscale,self.aprad/self.pixscale, num=10)
+        sbright= []
+        if not self.sky_global:
+            skyap = CircularAnnulus((obj['xcentroid'], obj['ycentroid']),
+                                    r_in=self.skyrad[0] / self.pixscale, 
+                                    r_out=self.skyrad[1] / self.pixscale)
+            skyphot = aperture_photometry(img, skyap)
+        for radius in radii:
+            ap = CircularAperture((obj['xcentroid'], obj['ycentroid']), radius)
+            if self.sky_global:
+                sbright.append( aperture_photometry(img - sky, ap)/ap.aera() )
+            else:
+                apphot = aperture_photometry(img, ap)
+                flux= apphot['aperture_sum'] - skyphot['aperture_sum'] / skyap.area() * ap.area()
+                sbright.append( flux/ap.area() )
+        # Sky subtracted surface brightness (nstars,napertures)
+        surfb= np.zeros( (len(sbright[0].data),len(radii)) )
+        for cnt in range(len(radii)):
+            surfb[:,cnt]= sbright[cnt].data
+        del sbright
+        # 20 brightest or the number left
+        nbright= min(20,len(obj))
+        ibright= np.argsort(surfb[:,0])[::-1][:nbright]
+        surfb= surfb[ibright,:]
+        # Non-linear least squares LM fit to Moffat Profile
+        fwhm= np.zeros(surfb.shape[0])
+        if not self.verboseplots: 
+            for cnt in range(surfb.shape[0]):
+                popt, pcov = curve_fit(moffatPSF, radii, surfb[cnt,:], p0 = [1.5*surfb[cnt,0], 5.,2.])
+                fwhm[cnt]= popt[1]
+        else: 
+            plt.close()
+            for cnt in range(surfb.shape[0]):
+                popt, pcov = curve_fit(moffatPSF, radii, surfb[cnt,:], p0 = [1.5*surfb[cnt,0], 5.,2.])
+                fwhm[cnt]= popt[1]
+                plt.plot(radii,surfb[cnt,:],'ok')
+                plt.plot(np.linspace(0,14,num=20),moffatPSF(np.linspace(0,14,num=20), *popt))
+            plt.xlabel('pixels')
+            fn= self.zptsfile.replace('.fits','_qa_fwhm_ccd%s.png' % ccds['image_hdu'].data[0])
+            plt.savefig(fn)
+            plt.close()
+            print('Wrote %s' % fn)
+        ccds['fwhm']= np.median(fwhm) * self.pixscale # arcsec
+        t0= ptime('fwhm-calculation',t0)
+        
+        # Now match against (good) PS1 stars 
+        # John cuts to magnitudes between 15 and 22
+        ps1 = ps1cat(ccdwcs=self.wcs).get_stars() #magrange=(15, 22))
+        good = (ps1.nmag_ok[:, 0] > 0)*(ps1.nmag_ok[:, 1] > 0)*(ps1.nmag_ok[:, 2] > 0)
+        # Get Gaia ra,dec
+        gdec=ps1.dec_ok-ps1.ddec/3600000.
+        gra=ps1.ra_ok-ps1.dra/3600000./np.cos(np.deg2rad(gdec))
+        # Cut 0.5 deg from CCD center and non star colors
+        gaia_cat = SkyCoord(ra=gra*units.degree, dec=gdec*units.degree)
+        center_ccd = SkyCoord(ra=ccds['ra']*units.degree, dec=ccds['dec']*units.degree)
+        ang = gaia_cat.separation(center_ccd) 
+        gicolor= ps1.median[:,0] - ps1.median[:,2]
+        good*= (np.array(ang) < 0.50)*(gicolor > 0.4)*(gicolor < 2.7)
+        # final cut
+        good = np.where(good)[0]
         ps1.cut(good)
+        gdec=ps1.dec_ok-ps1.ddec/3600000.
+        gra=ps1.ra_ok-ps1.dra/3600000./np.cos(np.deg2rad(gdec))
         nps1 = len(ps1)
 
         if nps1 == 0:
             print('No overlapping PS1 stars in this field!')
             return ccds, _stars_table()
-
-        # Compute the Gaia RA and DEC for the PS1 stars
-        gdec=ps1.dec_ok-ps1.ddec/3600000.
-        gra=ps1.ra_ok-ps1.dra/3600000./np.cos(np.deg2rad(gdec))
-        
-        objra, objdec = self.wcs.pixelxy2radec(obj['xcentroid']+1, obj['ycentroid']+1)
-        #m1, m2, d12 = match_radec(objra, objdec, ps1.ra, ps1.dec, self.matchradius/3600.0)
+    
+        # Match GAIA and Our Data
         m1, m2, d12 = match_radec(objra, objdec, gra, gdec, self.matchradius/3600.0)
         nmatch = len(m1)
         ccds['nmatch'] = nmatch
-        
-        #print('{} PS1 stars match detected sources within {} arcsec.'.format(nmatch, self.matchradius))
         print('{} GAIA sources match detected sources within {} arcsec.'.format(nmatch, self.matchradius))
         t0= ptime('match-to-gaia-radec',t0)
 
-        # Initialize the stars table and begin populating it.
+        # Stars table 
         print('Add the amplifier number!!!')
         stars = _stars_table(nmatch)
-        stars['filter'] = self.band
+        stars['image_filename'] =ccds['image_filename']
         stars['expid'] = self.expid
+        stars['filter'] = self.band
+        # Matched quantities
+        stars['nmatch'] = nmatch 
         stars['x'] = obj['xcentroid'][m1]
         stars['y'] = obj['ycentroid'][m1]
         stars['ra'] = objra[m1]
         stars['dec'] = objdec[m1]
-
-        apflux = apflux[m1] # we need apflux for Tractor, below
-        stars['apmag'] = - 2.5 * np.log10(apflux) + zp0 + 2.5 * np.log10(exptime)
-
-        #stars['ps1_ra'] = ps1.ra[m2]
-        #stars['ps1_dec'] = ps1.dec[m2]
-        stars['gaia_ra'] = gra[m2]
-        stars['gaia_dec'] = gdec[m2]
-        stars['ps1_gicolor'] = ps1.median[m2, 0] - ps1.median[m2, 2]
-
-        ps1band = ps1cat.ps1band[self.band]
-        stars['ps1_mag'] = ps1.median[m2, ps1band]
-
-        #plt.scatter(stars['ra'], stars['dec'], color='orange') ; plt.scatter(stars['ps1_ra'], stars['ps1_dec'], color='blue') ; plt.show()
-        
+        stars['radiff'] = (gra[m2] - stars['ra']) * np.cos(np.deg2rad(stars['dec'])) * 3600.0
+        stars['decdiff'] = (gdec[m2] - stars['dec']) * 3600.0
+        stars['apmag'] = - 2.5 * np.log10(apflux[m1]) + zp0 + 2.5 * np.log10(exptime)
+        # Add ps1 astrometric residuals for comparison
+        ps1_m1, ps1_m2, ps1_d12 = match_radec(objra, objdec, ps1.ra, ps1.dec, self.matchradius/3600.0)
+        stars['radiff_ps1'] = (ps1.ra[ps1_m2] - objra[ps1_m1]) * np.cos(np.deg2rad(objdec[ps1_m1])) * 3600.0
+        stars['decdiff_ps1'] = (ps1.dec[ps1_m2] - objdec[ps1_m1]) * 3600.0
+        # Photometry
         # Unless we're calibrating the photometric transformation, bring PS1
         # onto the photometric system of this camera (we add the color term
         # below).
         if self.calibrate:
+            raise ValueError('not Calibrating PS1 to our Camera, are you sure?')
             colorterm = np.zeros(nmatch)
         else:
             colorterm = self.colorterm_ps1_to_observed(ps1.median[m2, :], self.band)
-
-        # Compute the astrometric residuals relative to PS1.
-        #radiff = (stars['ra'] - stars['ps1_ra']) * np.cos(np.deg2rad(ccddec)) * 3600.0
-        #decdiff = (stars['dec'] - stars['ps1_dec']) * 3600.0
-        radiff = (stars['ra'] - stars['gaia_ra']) * np.cos(np.deg2rad(ccddec)) * 3600.0
-        decdiff = (stars['dec'] - stars['gaia_dec']) * 3600.0
-        ccds['raoff'] = np.median(radiff)
-        ccds['decoff'] = np.median(decdiff)
-        ccds['rarms'] = np.std(radiff)
-        ccds['decrms'] = np.std(decdiff)
-        #print('RA, Dec offsets (arcsec) relative to PS1: {}, {}'.format(ccds['raoff'], ccds['decoff']))
-        #print('RA, Dec rms (arcsec) relative to PS1: {}, {}'.format(ccds['rarms'], ccds['decrms']))
-        print('RA, Dec offsets (arcsec) relative to GAIA: {}, {}'.format(ccds['raoff'], ccds['decoff']))
-        print('RA, Dec rms (arcsec) relative to GAIA: {}, {}'.format(ccds['rarms'], ccds['decrms']))
-
-        # Compute the photometric zeropoint but only use stars with main
-        # sequence g-i colors.
+        ps1band = ps1cat.ps1band[self.band]
+        stars['ps1_mag'] = ps1.median[m2, ps1band] + colorterm
+        # Additonal mags for comparison with Arjun's star sample
+        # PS1 Median PSF mag in [g,r,i,z],  Gaia G-band mean magnitude
+        for ps1_band,ps1_index in zip(['g','r','i','z'],[0,1,2,3]):
+            stars['ps1_%s' % ps1_band]= ps1.median[m2, ps1_index]
+        stars['gaia_g']=ps1.phot_g_mean_mag[m2]
+        
+        # Zeropoint Sample is Main Sequence stars
         print('Computing the photometric zeropoint.')
+        stars['ps1_gicolor'] = ps1.median[m2, 0] - ps1.median[m2, 2]
+        print('Before gicolor cut, len(stars)=%d' % len(stars['ps1_gicolor']))
         mskeep = np.where((stars['ps1_gicolor'] > 0.4) * (stars['ps1_gicolor'] < 2.7))[0]
         if len(mskeep) == 0:
             print('Not enough PS1 stars with main sequence colors.')
             return ccds, stars
         ccds['mdncol'] = np.median(stars['ps1_gicolor'][mskeep]) # median g-i color
-
-        # Get the photometric offset relative to PS1 as the observed PS1
-        # magnitude minus the observed / measured magnitude.
-
-        stars['ps1_mag'] += colorterm
-        #plt.scatter(stars['ps1_gicolor'], stars['apmag']-stars['ps1_mag']) ; plt.show()
+        print('After gicolor cut, len(stars)=%d' % len(mskeep))
+        
+        # Compute Zeropoint
         dmagall = stars['ps1_mag'][mskeep] - stars['apmag'][mskeep]
-        _, dmagsig = self.sensible_sigmaclip(dmagall, nsigma=2.5)
-
-        dmag, _, _ = sigmaclip(dmagall, low=3, high=3.0)
+        dmag, _, _ = sigmaclip(dmagall, low=2.5, high=2.5)
         dmagmed = np.median(dmag)
         ndmag = len(dmag)
+        # Std dev
+        #_, dmagsig = self.sensible_sigmaclip(dmagall, nsigma=2.5)
+        dmagsig = np.std(dmag)  # agrees with IDL codes, they just compute std
 
         zptmed = zp0 + dmagmed
         transp = 10.**(-0.4 * (zp0 - zptmed - kext * (airmass - 1.0)))
 
-        print('  Mag offset: {:.3f}'.format(dmagmed))
-        print('  Scatter:    {:.3f}'.format(dmagsig))
-        
-        print('  {} stars used for zeropoint median'.format(ndmag))
-        print('  Zeropoint {:.3f}'.format(zptmed))
-        print('  Transparency: {:.3f}'.format(transp))
         t0= ptime('photometry-using-ps1',t0)
-
+        ccds['raoff'] = np.median(stars['radiff'])
+        ccds['decoff'] = np.median(stars['decdiff'])
+        ccds['rarms'] = np.std(stars['radiff'])
+        ccds['decrms'] = np.std(stars['decdiff'])
         ccds['phoff'] = dmagmed
         ccds['phrms'] = dmagsig
         ccds['zpt'] = zptmed
         ccds['transp'] = transp
 
+        print('RA, Dec offsets (arcsec) relative to GAIA: {}, {}'.format(ccds['raoff'], ccds['decoff']))
+        print('RA, Dec rms (arcsec) relative to GAIA: {}, {}'.format(ccds['rarms'], ccds['decrms']))
+        print('  Mag offset: {}'.format(ccds['phoff']))
+        print('  Scatter:    {}'.format(ccds['phrms']))
+        
+        print('  {} stars used for zeropoint median'.format(ndmag))
+        print('  Zeropoint {}'.format(ccds['zpt']))
+        print('  Transparency: {}'.format(ccds['transp']))
+
+        t0= ptime('all-computations-for-this-ccd',t0)
+        # Plots for comparing to Arjuns zeropoints*.ps
+        self.make_plots(stars,dmag,ccds['zpt'],ccds['transp'])
+        t0= ptime('made-plots',t0)
+
+        # No longer neeeded: 
         # Fit each star with Tractor.
         # Skip for now, most time consuming part
         #ivar = np.zeros_like(img) + 1.0/sig1**2
@@ -689,13 +823,69 @@ class Measurer(object):
         
         return ccds, stars
     
+    def make_plots(self,stars,dmag,zpt,transp):
+        '''stars -- stars table'''
+        suffix='_qa_%s.png' % stars['expid'][0][-4:]
+        fig,ax=plt.subplots(1,2,figsize=(10,4))
+        plt.subplots_adjust(wspace=0.2,bottom=0.2,right=0.8)
+        for key in ['astrom_gaia','photom']:
+        #for key in ['astrom_gaia','astrom_ps1','photom']:
+            if key == 'astrom_gaia':    
+                ax[0].scatter(stars['radiff'],stars['decdiff'])
+                xlab=ax[0].set_xlabel(r'$\Delta Ra$ (Gaia - CCD)')
+                ylab=ax[0].set_ylabel(r'$\Delta Dec$ (Gaia - CCD)')
+            elif key == 'astrom_ps1':  
+                raise ValueError('not needed')  
+                ax.scatter(stars['radiff_ps1'],stars['decdiff_ps1'])
+                ax.set_xlabel(r'$\Delta Ra [arcsec]$ (PS1 - CCD)')
+                ax.set_ylabel(r'$\Delta Dec [arcsec]$ (PS1 - CCD)')
+                ax.text(0.02, 0.95,'Median: %.4f,%.4f' % \
+                          (np.median(stars['radiff_ps1']),np.median(stars['decdiff_ps1'])),\
+                        va='center',ha='left',transform=ax.transAxes,fontsize=20)
+                ax.text(0.02, 0.85,'RMS: %.4f,%.4f' % \
+                          (getrms(stars['radiff_ps1']),getrms(stars['decdiff_ps1'])),\
+                        va='center',ha='left',transform=ax.transAxes,fontsize=20)
+            elif key == 'photom':
+                ax[1].hist(dmag)
+                xlab=ax[1].set_xlabel('PS1 - AP mag (main seq, 2.5 clipping)')
+                ylab=ax[1].set_ylabel('Number of Stars')
+        # List key numbers
+        ax[1].text(1.02, 1.,r'$\Delta$ Ra,Dec',\
+                va='center',ha='left',transform=ax[1].transAxes,fontsize=12)
+        ax[1].text(1.02, 0.9,r'  Median: %.4f,%.4f' % \
+                  (np.median(stars['radiff']),np.median(stars['decdiff'])),\
+                va='center',ha='left',transform=ax[1].transAxes,fontsize=10)
+        ax[1].text(1.02, 0.80,'  RMS: %.4f,%.4f' % \
+                  (getrms(stars['radiff']),getrms(stars['decdiff'])),\
+                va='center',ha='left',transform=ax[1].transAxes,fontsize=10)
+        ax[1].text(1.02, 0.7,'PS1-CCD Mag',\
+                va='center',ha='left',transform=ax[1].transAxes,fontsize=12)
+        ax[1].text(1.02, 0.6,'  Median:%.4f,%.4f' % \
+                  (np.median(dmag),np.std(dmag)),\
+                va='center',ha='left',transform=ax[1].transAxes,fontsize=10)
+        ax[1].text(1.02, 0.5,'  Stars: %d' % len(dmag),\
+                va='center',ha='left',transform=ax[1].transAxes,fontsize=10)
+        ax[1].text(1.02, 0.4,'  Zpt=%.4f' % zpt,\
+                va='center',ha='left',transform=ax[1].transAxes,fontsize=10)
+        ax[1].text(1.02, 0.3,'  Transp=%.4f' % transp,\
+                va='center',ha='left',transform=ax[1].transAxes,fontsize=10)
+        # Save
+        fn= self.zptsfile.replace('.fits',suffix)
+        plt.savefig(fn,bbox_extra_artists=[xlab,ylab])
+        plt.close()
+        print('Wrote %s' % fn)
+   
+ 
 class DecamMeasurer(Measurer):
-    '''Class to measure a variety of quantities from a single DECam CCD.'''
+    '''Class to measure a variety of quantities from a single DECam CCD.
+    UNITS: ADU/s'''
     def __init__(self, *args, **kwargs):
         super(DecamMeasurer, self).__init__(*args, **kwargs)
 
+        self.pixscale=0.262 #fixed
         self.camera = 'decam'
         self.ut = self.primhdr['TIME-OBS']
+        self.band = self.get_band()
         self.ra_bore = hmsstring2ra(self.primhdr['TELRA'])
         self.dec_bore = dmsstring2dec(self.primhdr['TELDEC'])
         self.gain = self.hdr['ARAWGAIN'] # hack! average gain [electron/sec]
@@ -725,20 +915,28 @@ class DecamMeasurer(Measurer):
         img *= self.gain
         #img *= self.gain / self.exptime
         return img, hdr
+     
+    def get_wcs(self):
+        return wcs_pv2sip_hdr(self.hdr) # PV distortion
 
 class Mosaic3Measurer(Measurer):
-    '''Class to measure a variety of quantities from a single Mosaic3 CCD.'''
+    '''Class to measure a variety of quantities from a single Mosaic3 CCD.
+    UNITS: e-/s'''
     def __init__(self, *args, **kwargs):
         super(Mosaic3Measurer, self).__init__(*args, **kwargs)
 
+        self.pixscale=0.260 #KJB, see mosstat.pro
         self.camera = 'mosaic3'
+        self.band= self.get_band()
         self.ut = self.primhdr['TIME-OBS']
         self.ra_bore = hmsstring2ra(self.primhdr['TELRA'])
         self.dec_bore = dmsstring2dec(self.primhdr['TELDEC'])
-        self.gain = self.hdr['GAIN'] # hack! average gain
+        #self.gain = self.hdr['GAIN'] # hack! average gain
+        self.gain = 1.8 #Average raw, see mosstat
 
         print('Hack! Using an average Mosaic3 zeropoint!!')
-        corr = 2.5 * np.log10(self.gain)
+        #corr = 2.5 * np.log10(self.gain) 
+        corr = 0. 
         self.zp0 = dict(z = 26.552 + corr)
         self.sky0 = dict(z = 18.46 + corr)
         self.k_ext = dict(z = 0.06)
@@ -758,16 +956,26 @@ class Mosaic3Measurer(Measurer):
         #fits=fitsio.FITS(fn,mode='r',clobber=False,lower=True)
         #hdr= fits[0].read_header()
         #img= fits[ext].read()
-        #img *= self.exptime
-        img *= self.exptime*self.gain #KJB
+        img *= self.exptime 
         return img, hdr
 
+    def read_bitmask(self):
+        fn= self.fn.replace('ooi','ood')
+        mask, junk = fitsio.read(fn, ext=self.ext, header=True)
+        return mask
+    
+    def get_wcs(self):
+        return wcs_pv2sip_hdr(self.hdr) # PV distortion
+
 class NinetyPrimeMeasurer(Measurer):
-    '''Class to measure a variety of quantities from a single 90prime CCD.'''
+    '''Class to measure a variety of quantities from a single 90prime CCD.
+    UNITS -- CP e-/s'''
     def __init__(self, *args, **kwargs):
         super(NinetyPrimeMeasurer, self).__init__(*args, **kwargs)
         
+        self.pixscale=0.455 #KJB, fixed
         self.camera = '90prime'
+        self.band= self.get_band()
         self.ra_bore = hmsstring2ra(self.primhdr['RA'])
         self.dec_bore = dmsstring2dec(self.primhdr['DEC'])
         self.ut = self.primhdr['UT']
@@ -776,12 +984,14 @@ class NinetyPrimeMeasurer(Measurer):
         # information should be scraped from the headers, plus we're ignoring
         # the gain variations across amplifiers (on a given CCD).
         gaindict = dict(ccd1 = 1.47, ccd2 = 1.48, ccd3 = 1.42, ccd4 = 1.4275)
-        self.gain = gaindict[self.ccdname.lower()]
+        #self.gain = gaindict[self.ccdname.lower()]
+        self.gain = 1.3 #KJB, average raw see bokstat
 
         # Nominal zeropoints, sky brightness, and extinction values (taken from
         # rapala.ninetyprime.boketc.py).  The sky and zeropoints are both in
         # ADU, so account for the gain here.
-        corr = 2.5 * np.log10(self.gain)
+        #corr = 2.5 * np.log10(self.gain)
+        corr=0.
         self.zp0 = dict(g = 25.55 + corr, r = 25.38 + corr)
         self.sky0 = dict(g = 22.10 + corr, r = 21.07 + corr)
         self.k_ext = dict(g = 0.17, r = 0.10)
@@ -789,8 +999,7 @@ class NinetyPrimeMeasurer(Measurer):
     def get_band(self):
         band = self.primhdr['FILTER']
         band = band.split()[0]
-        band.replace('bokr', 'r')
-        return band
+        return band.replace('bokr', 'r')
 
     def colorterm_ps1_to_observed(self, ps1stars, band):
         from legacyanalysis.ps1cat import ps1_to_90prime
@@ -799,12 +1008,18 @@ class NinetyPrimeMeasurer(Measurer):
     def read_image(self):
         '''Read the image and header.  Convert image from electrons/sec to electrons.'''
         img, hdr = fitsio.read(self.fn, ext=self.ext, header=True)
-        #fits=fitsio.FITS(fn,mode='r',clobber=False,lower=True)
-        #hdr= fits[0].read_header()
-        #img= fits[ext].read()
         img *= self.exptime
         return img, hdr
+    
+    def read_bitmask(self):
+        fn= self.fn.replace('ooi','ood')
+        mask, junk = fitsio.read(fn, ext=self.ext, header=True)
+        return mask
   
+    def get_wcs(self):
+        return wcs_pv2sip_hdr(self.hdr) # PV distortion
+
+
 def camera_name(primhdr):
     '''
     Returns 'mosaic3', 'decam', or '90prime'
@@ -850,7 +1065,7 @@ def _measure_image(args):
     '''Utility function to wrap measure_image function for multiprocessing map.''' 
     return measure_image(*args)
 
-def measure_image(img_fn, measureargs={}):
+def measure_image(img_fn, **measureargs): 
     '''Wrapper on the camera-specific classes to measure the CCD-level data on all
     the FITS extensions for a given set of images.
     '''
@@ -887,40 +1102,34 @@ def measure_image(img_fn, measureargs={}):
     return ccds, stars
 
 
-def get_output_fns(img_fn):
+def get_output_fns(img_fn,prefix=''):
     zptsfile= os.path.dirname(img_fn).replace('/project/projectdirs','/scratch2/scratchdirs/kaylanb')
-    zptsfile= os.path.join(zptsfile,'zpts/','only_ccd_centers_zeropoint-%s' % os.path.basename(img_fn))
+    zptsfile= os.path.join(zptsfile,'zpts/','%szeropoint-%s' % (prefix,os.path.basename(img_fn)))
     zptsfile= zptsfile.replace('.fz','')
     zptstarsfile = zptsfile.replace('.fits','-stars.fits')
-    #zptsfile = os.path.join(outdir, '{}.fits'.format(prefix))
-    #zptstarsfile = os.path.join(outdir, '{}-stars.fits'.format(prefix))
     return zptsfile,zptstarsfile
 
 
-def runit(img_fn,measureargs, zptsfile='zpt.fits',zptstarsfile='zptstar.fits'):
+def runit(img_fn, **measureargs):
     '''Generate a legacypipe-compatible CCDs file for a given image.
 
     '''
+    zptsfile= measureargs.get('zptsfile')
+    zptstarsfile= measureargs.get('zptstarsfile')
+
     t0 = Time()
     if not os.path.exists(os.path.dirname(zptsfile)):
         dobash('mkdir -p %s' % os.path.dirname(zptsfile))
-        #os.makedirs(os.path.dirname(zptsfile))
 
     # Copy to SCRATCH for improved I/O
     fn_scr= img_fn.replace('/project/projectdirs','/scratch2/scratchdirs/kaylanb')
     if not os.path.exists(fn_scr): 
         dobash("cp %s %s" % (img_fn,fn_scr))
-        t0= ptime('copy-to-scratch',t0)
+    if not os.path.exists(fn_scr.replace('_ooi_','_ood_')): 
+        dobash("cp %s %s" % (img_fn.replace('_ooi_','_ood_'),fn_scr.replace('_ooi_','_ood_')))
+    t0= ptime('copy-to-scratch',t0)
 
-    ccds, stars= measure_image(fn_scr, measureargs)
-    # If combining ccds from multiple images:
-    #allccds = []
-    #if len(allccds) == 0:
-    #    allccds = ccds
-    #    allstars = stars
-    #else:
-    #    allccds = vstack((allccds, ccds))
-    #    allstars = vstack((allstars, stars))
+    ccds, stars= measure_image(fn_scr, **measureargs)
     t0= ptime('measure_image',t0)
 
     # Write out.
@@ -935,17 +1144,317 @@ def runit(img_fn,measureargs, zptsfile='zpt.fits',zptstarsfile='zptstar.fits'):
     if os.path.exists(fn_scr): 
         assert(fn_scr.startswith('/scratch2/scratchdirs/kaylanb'))
         dobash("rm %s" % fn_scr)
+        dobash("rm %s" % fn_scr.replace('_ooi_','_ood_'))
         t0= ptime('removed-cp-from-scratch',t0)
     
-                    
+class Compare2Arjuns(object):
+    '''contains the functions to compare every column of legacy ccd_table to 
+    that of Arjun's zeropoints files'''
+    def __init__(self,zptfn_list): 
+        '''combines many zpt files into one table, 
+        does this for legacy and the corresponding zpt tables from Arjun
+        givent the relative path to arjun's zpt tables
+        
+        zptfn_list: text file listing each legacy zpt file to be used
+        camera: ['mosaic','90prime','decam']
+        '''
+        camera= self.get_camera(zptfn_list)
+         
+        if camera == 'mosaic':
+            self.path_to_arjuns= '/scratch2/scratchdirs/arjundey/ZeroPoints_MzLSv2'
+        elif camera == '90prime':
+            self.path_to_arjuns= '/scratch2/scratchdirs/arjundey/ZeroPoints_BASS'
+
+        # Get legacy zeropoints, and corresponding ones from Arjun
+        self.makeBigTable(zptfn_list)
+        self.ccd_cuts(camera)
+        # Compare values
+        self.getKeyTypes()
+        self.compare_alphabetic()
+        self.compare_numeric()
+        self.compare_numeric_stars()
+        raise ValueError
+
+    def get_camera(self,zptfn_list):
+        fns= np.loadtxt(zptfn_list,dtype=str)
+        if fns.size == 1:
+            fns= [str(fns)]
+        fn=fns[0]
+        
+        if 'ksb' in fn:
+            camera='90prime'
+        elif 'k4m' in fn:
+            camera= 'mosaic'
+        elif 'c4d' in fn:
+            camera= 'decam'
+        else: raise ValueError('camera not clear from fn=%s' % tempfn)
+        return camera
+
+    def makeBigTable(self,zptfn_list):
+        '''combines many zpt files into one table, 
+        does this for legacy and the corresponding zpt tables from Arjun
+        givent the relative path to arjun's zpt tables
+        
+        zptfn_list: text file listing each legacy zpt file to be used
+        '''
+        self.legacy,self.legacy_stars,self.arjun,self.arjun_stars= [],[],[],[]
+        # Simultaneously read in arjun's with legacy
+        fns= np.loadtxt(zptfn_list,dtype=str)
+        if fns.size == 1:
+            fns= [str(fns)]
+        for cnt,fn in enumerate(fns):
+            print('%d/%d: ' % (cnt+1,len(fns)))
+            try:
+                # Legacy zeropoints, use Arjun's naming scheme
+                self.legacy.append( self.read_legacy(fn,reset_names=True) ) 
+                fn_stars= fn.replace('.fits','-stars.fits')
+                self.legacy_stars.append( self.read_legacy(fn_stars,reset_names=True,stars=True) )
+                # Corresponding zeropoints from Arjun
+                arjun_fn= os.path.basename(fn)
+                index= arjun_fn.find('zeropoint') # Check for a prefix
+                if index > 0: arjun_fn= arjun_fn.replace(arjun_fn[:index],'')
+                arjun_fn= os.path.join(self.path_to_arjuns, arjun_fn)
+                self.arjun.append( fits_table(arjun_fn) ) 
+                self.arjun_stars.append( fits_table(arjun_fn.replace('zeropoint-','matches-') ))
+            except IOError:
+                print('WARNING: one of these cannot be read: %s\n%s\n%s\n%s\n' % \
+                     (fn,fn.replace('.fits','-stars.fits'),\
+                      arjun_fn,arjun_fn.replace('zeropoint-','matches-'))
+                     )
+        self.legacy= merge_tables(self.legacy, columns='fillzero') 
+        self.legacy_stars= merge_tables(self.legacy_stars, columns='fillzero') 
+        self.arjun= merge_tables(self.arjun, columns='fillzero') 
+        self.arjun_stars= merge_tables(self.arjun_stars, columns='fillzero')
+    
+    def ccd_cuts(self,camera):
+        keep= np.zeros(len(self.legacy)).astype(bool)
+        for tab in [self.legacy,self.arjun]:
+            if camera == 'mosaic':
+                keep[ (tab.exptime > 40.)*(tab.ccdnmatch > 50)*(tab.ccdzpt > 25.8) ] = True
+            elif camera == '90prime':
+                keep[ (tab.ccdzpt >= 20.)*(tab.ccdzpt <= 30.) ] = True
+        self.legacy.cut(keep)
+        self.arjun.cut(keep)
+
+
+    def read_legacy(self,zptfn,reset_names=True,stars=False):
+        '''reads in a legacy zeropoint table as a fits_table() object
+        reset_names: 
+            True -- rename everything to give it Arjun's naming scheme
+            False -- return table as is
+        stars:
+            True -- the input is the stars table that accompanies each
+                    zeropoints table, e.g. instead of the zpt table
+            False -- the input is the zeropoitn table
+        '''
+        if stars:
+            # Just columns we'll compare
+            translate= dict(ra='ccd_ra',\
+                            dec='ccd_dec',\
+                            apmag='ccd_mag',\
+                            radiff='raoff',\
+                            decdiff='decoff',\
+                            gaia_g='gmag')
+            self.star_keys= []
+            for key in translate.keys():
+                self.star_keys.append( translate[key] ) 
+            # Add ps1_g,r... to compare those too
+            for band in ['g','r','i','z']:
+                self.star_keys.append( 'ps1_%s' % band )
+        else:
+            legacy_missing= ['ccdhdu','seeing',\
+                           'ccdnmatcha','ccdnmatchb','ccdnmatchc','ccdnmatchd',\
+                           'ccdzpta','ccdzptb','ccdzptc','ccdzptd',\
+                           'ccdnum',\
+                           'psfab','psfpa','temp','badimg']
+            
+            arjun_missing= ['camera','expid','pixscale']
+            self.missing= legacy_missing + arjun_missing
+           
+            # All columns without matching name
+            translate= dict(image_filename='filename',\
+                            image_hdu='ccdhdunum',\
+                            gain='arawgain',\
+                            width='naxis1',\
+                            height='naxis2',\
+                            ra='ccdra',\
+                            dec='ccddec',\
+                            ra_bore='ra',\
+                            dec_bore='dec',\
+                            raoff='ccdraoff',\
+                            decoff='ccddecoff',\
+                            rarms='ccdrarms',\
+                            decrms='ccddecrms',\
+                            skycounts='ccdskycounts',\
+                            skymag='ccdskymag',\
+                            skyrms='ccdskyrms',\
+                            nstar='ccdnstar',\
+                            nmatch='ccdnmatch',\
+                            mdncol='ccdmdncol',\
+                            phoff='ccdphoff',\
+                            phrms='ccdphrms',\
+                            transp='ccdtransp',\
+                            zpt='ccdzpt',\
+                            zptavg='zpt')
+        
+        legacy=fits_table(zptfn)
+        if reset_names:
+            for key in translate.keys():
+                legacy.rename(key, translate[key]) #(old,new)
+        return legacy
+
+
+
+    def getKeyTypes(self):
+        '''sorts keys as either numeric or alphabetic'''
+        self.numeric_keys=[]
+        self.alpha_keys=[]
+        for key in self.legacy.get_columns():
+            if key in self.missing:
+                continue # Either not in legacy or not in Arjuns
+            typ= type(self.legacy.get(key)[0])
+            if np.any((typ == np.float32,\
+                       typ == np.float64),axis=0):
+                self.numeric_keys+= [key]
+            elif np.any((typ == np.int16,\
+                         typ == np.int32),axis=0):
+                self.numeric_keys+= [key]
+            elif typ == np.string_:
+                self.alpha_keys+= [key]
+            else:
+                print('WARNING: unknown type for key=%s, ' % key,typ)
+
+    def compare_alphabetic(self):
+        print('-'*20)
+        print('legacy == arjuns:')
+        for key in self.alpha_keys:
+            # Simply compare first row only, not all rows
+            if key in ['filename']:
+                print('%s: ' % key,self.legacy.get(key)[0].replace('.fits.fz','.fits') == self.arjun.get(key)[0])
+            else:
+                print('%s: ' % key,self.legacy.get(key)[0] == self.arjun.get(key)[0])
+
+    def compare_numeric(self):
+        '''two plots of everything numberic between legacy zeropoints and Arjun's
+        1) x vs. y 
+        2) x vs. (y-x)/|y+x|
+        '''
+        for doplot in ['dpercent','default']:
+            panels=len(self.numeric_keys)
+            cols=3
+            if panels % cols == 0:
+                rows=panels/cols
+            else:
+                rows=panels/cols+1
+            rows=int(rows)
+            fig,axes= plt.subplots(rows,cols,figsize=(20,30))
+            ax=axes.flatten()
+            plt.subplots_adjust(hspace=0.4,wspace=0.3)
+            xlims,ylims= None,None
+            for cnt,key in enumerate(self.numeric_keys):
+                x= self.arjun.get(key)
+                ti= 'Labels: x-axis = A, '
+                if doplot == 'dpercent':
+                    y= ( self.arjun.get(key) - self.legacy.get(key) ) / \
+                       np.abs( self.arjun.get(key) + self.legacy.get(key) )
+                    ti+= ' y-axis = (A - L)/|A + L|'
+                    ylims=[-0.1,0.1]
+                elif doplot == 'default':
+                    y= self.legacy.get(key)
+                    ti+= ' y-axis = L'
+                    xlims= [ min([x.min(),y.min()]),max([x.max(),y.max()]) ]
+                    if xlims[0] < 0: xlims[0]*=1.02
+                    else: xlims[0]*=0.98
+                    if xlims[1] < 0: xlims[0]*=0.98
+                    else: xlims[1]*=1.02
+                    ylims= xlims
+                else: raise ValueError('%s not allowed' % doplot)
+                ax[cnt].scatter(x,y) 
+                ax[cnt].text(0.025,0.88,key,\
+                             va='center',ha='left',transform=ax[cnt].transAxes,fontsize=20) 
+                if xlims is not None:
+                    ax[cnt].set_xlim(xlims)
+                if ylims is not None:
+                    ax[cnt].set_ylim(ylims)
+            ax[1].text(0.5,1.5,ti,\
+                       va='center',ha='center',transform=ax[1].transAxes,fontsize=30)
+            fn="numeric_%s.png" % doplot
+            plt.savefig(fn) 
+            print('Wrote %s' % fn)
+            plt.close()
+    
+    def compare_numeric_stars(self):
+        '''two plots of everything numberic between legacy Stars and Arjun's Stars
+        1) x vs. y 
+        2) x vs. (y-x)/|y+x|
+        '''
+        # Match
+        m1, m2, d12 = match_radec(self.arjun_stars.ccd_ra, self.arjun_stars.ccd_dec, \
+                                  self.legacy_stars.ccd_ra, self.legacy_stars.ccd_dec, \
+                                1./3600.0)
+        nstars=dict(arjun=len(self.arjun_stars),legacy=len(self.legacy_stars))
+        self.arjun_stars.cut(m1)
+        self.legacy_stars.cut(m2)
+        assert(len(self.arjun_stars) == len(self.legacy_stars))
+        ti_top= 'Matched Stars:%d, Arjun had:%d, Legacy had:%d' % \
+                (len(self.legacy_stars),nstars['arjun'],nstars['legacy'])
+        # Plot
+        for doplot in ['dpercent','default']:
+            panels=len(self.star_keys)
+            cols=3
+            if panels % cols == 0:
+                rows=panels/cols
+            else:
+                rows=panels/cols+1
+            rows=int(rows)
+            fig,axes= plt.subplots(rows,cols,figsize=(20,10))
+            ax=axes.flatten()
+            plt.subplots_adjust(hspace=0.4,wspace=0.3)
+            xlims,ylims= None,None
+            for cnt,key in enumerate(self.star_keys):
+                x= self.arjun_stars.get(key)
+                ti= 'Labels: x-axis = A, '
+                if doplot == 'dpercent':
+                    y= ( self.arjun_stars.get(key) - self.legacy_stars.get(key) ) / \
+                       np.abs( self.arjun_stars.get(key) + self.legacy_stars.get(key) )
+                    ti+= ' y-axis = (A - L)/|A + L|'
+                    ylims=[-0.01,0.01]
+                elif doplot == 'default':
+                    y= self.legacy_stars.get(key)
+                    ti+= 'y-axis = L'
+                    xlims= [ min([x.min(),y.min()]),max([x.max(),y.max()]) ]
+                    if xlims[0] < 0: xlims[0]*=1.02
+                    else: xlims[0]*=0.98
+                    if xlims[1] < 0: xlims[0]*=0.98
+                    else: xlims[1]*=1.02
+                    ylims= xlims
+                else: raise ValueError('%s not allowed' % doplot)
+                ax[cnt].scatter(x,y) 
+                ax[cnt].text(0.025,0.88,key,\
+                             va='center',ha='left',transform=ax[cnt].transAxes,fontsize=20) 
+                if xlims is not None:
+                    ax[cnt].set_xlim(xlims)
+                if ylims is not None:
+                    ax[cnt].set_ylim(ylims)
+            ax[1].text(0.5,1.5,ti_top,\
+                       va='center',ha='center',transform=ax[1].transAxes,fontsize=20)
+            ax[1].text(0.5,1.3,ti,\
+                       va='center',ha='center',transform=ax[1].transAxes,fontsize=20)
+            fn="numeric_stars_%s.png" % doplot
+            plt.savefig(fn) 
+            print('Wrote %s' % fn)
+            plt.close()
+ 
+                
 if __name__ == "__main__":
     t0 = Time()
     tbegin=t0
     print('TIMING:after-imports ',datetime.datetime.now())
     parser = argparse.ArgumentParser(description='Generate a legacypipe-compatible CCDs file from a set of reduced imaging.')
-    parser.add_argument('--image_list',action='store',help='List of images to process',required=True)
-    parser.add_argument('--jobid',action='store',help='slurm jobid',default='001',required=False)
-    parser.add_argument('--prefix', type=str, default='zeropoints', help='Prefix to prepend to the output files.')
+    parser.add_argument('--image_list',action='store',help='List of images to process, if compare2arjun = True then list of legacy zeropoint files',required=True)
+    parser.add_argument('--prefix', type=str, default='', help='Prefix to prepend to the output files.')
+    parser.add_argument('--verboseplots', action='store_true', default=False, help='use to plot FWHM Moffat PSF fits to the 20 brightest stars')
+    parser.add_argument('--compare2arjun', action='store_true', default=False, help='turn this on and give --image-list a list of legacy zeropoint files instead of cp images')
     parser.add_argument('--outdir', type=str, default='./legacy_zpt_outdir', help='Output directory.')
     parser.add_argument('--aprad', type=float, default=3.5, help='Aperture photometry radius (arcsec).')
     parser.add_argument('--skyrad-inner', type=float, default=7.0, help='Radius of inner sky annulus (arcsec).')
@@ -955,50 +1464,52 @@ if __name__ == "__main__":
                         help='Use this option when deriving the photometric transformation equations.')
     parser.add_argument('--sky-global', action='store_true',
                         help='Use a global rather than a local sky-subtraction around the stars.')
-    parser.add_argument('--only_ccd_centers', action='store_true', default=False, help='get ccd ra,dec centers only')
 
     args = parser.parse_args()
-    
+   
+    if args.compare2arjun:
+        comp= Compare2Arjuns(args.image_list)
+        raise ValueError('Done')
+
+ 
     images= read_lines(args.image_list) 
-    #images= glob(args.images) 
     
     # Build a dictionary with the optional inputs.
     measureargs = vars(args)
+    if not args.compare2arjun:
+        measureargs.pop('compare2arjun')
     measureargs.pop('image_list')
-    #images = np.array(measureargs.pop('images'))
     nproc = measureargs.pop('nproc')
 
-    prefix = measureargs.pop('prefix')
     outdir = measureargs.pop('outdir')
     if not os.path.exists(outdir):
         os.makedirs(outdir)
-    jobid = measureargs.pop('jobid')
-
-    print("measureargs['only_ccd_centers']= ",measureargs['only_ccd_centers'])
 
     if nproc > 1:
         from mpi4py.MPI import COMM_WORLD as comm
     t0=ptime('parse-args',t0)
 
-    
-    # RUN HERE
+    # MPI4py
     if nproc > 1:
         images_split= np.array_split(images, comm.size)
         for image_fn in images_split[comm.rank]:
             # Check if zpt already written
-            zptsfile,zptstarsfile= get_output_fns(image_fn)
+            zptsfile,zptstarsfile= get_output_fns(image_fn,prefix=measureargs.get('prefix'))
             if os.path.exists(zptsfile) and os.path.exists(zptstarsfile):
                 print('Skipping b/c exists: %s' % zptsfile)
                 continue  # Already done
+            measureargs.update(dict(zptsfile=zptsfile,\
+                                    zptstarsfile=zptstarsfile))
             # Log to unique file
-            outfn=os.path.join(outdir,"std.zpt%s_jobid%s" % (os.path.basename(image_fn),jobid))  
+            outfn=os.path.join(outdir,"std.zpt-%s%s" % \
+                        (os.path.basename(image_fn),\
+                        datetime.datetime.now().strftime("m%m-d%d-hr%H-min%M")))  
             with stdouterr_redirected(to=outfn, comm=None):  
                 t0=ptime('b4-run',t0)
-                runit(image_fn, measureargs, \
-                      zptsfile=zptsfile,zptstarsfile=zptstarsfile)
+                runit(image_fn, **measureargs)
                 t0=ptime('after-run',t0)
                 # Finish up 
-        # HACK, not sure if need to wait for all proc to finish 
+        # Wait for all mpi tasks to finish 
         confirm_files = comm.gather( images_split[comm.rank], root=0 )
         if comm.rank == 0:
             print('Rank 0 gathered the results:')
@@ -1006,17 +1517,19 @@ if __name__ == "__main__":
             tnow= Time()
             print("TIMING:total %s" % (tnow-tbegin,))
             print("Done")
+    # Serial
     else:
         for image_fn in images:
             # Check if zpt already written
-            zptsfile,zptstarsfile= get_output_fns(image_fn)
+            zptsfile,zptstarsfile= get_output_fns(image_fn,prefix=measureargs.get('prefix'))
             if os.path.exists(zptsfile) and os.path.exists(zptstarsfile):
                 print('continuing'.upper())
                 continue  # Already done
+            measureargs.update(dict(zptsfile=zptsfile,\
+                                    zptstarsfile=zptstarsfile))
             # Create the file
             t0=ptime('b4-run',t0)
-            runit(image_fn, measureargs,\
-                  zptsfile=zptsfile,zptstarsfile=zptstarsfile)
+            runit(image_fn, **measureargs)
             t0=ptime('after-run',t0)
         tnow= Time()
         print("TIMING:total %s" % (tnow-tbegin,))
